Fixes warnings when compiled with wxWidgets 2.8+
svn path=/trunk/flamerobin/; revision=1301

+++ src/gui/AdvancedSearchFrame.cpp
@@ -648,7 +648,7 @@ void AdvancedSearchFrame::OnButtonAddDatabaseClick(wxCommandEvent& WXUNUSED(even
 {
     if (choice_database->GetSelection() == 0)   // all connected databases
     {
-        for (int i=1; i<choice_database->GetCount(); ++i)
+        for (int i = 1; i < (int)choice_database->GetCount(); ++i)
         {
             Database *db = (Database *)choice_database->GetClientData(i);
             if (db && db->isConnected())
